Fix int value wrong bug (should use sbyte)

+++ Tjs2Net/Engine/TjsByteCodeLoader.cs
@@ -406,7 +406,7 @@ namespace Tjs2
 
                         case TYPE_BYTE:
                         {
-                            vdata[i] = new Variant(barray[index]);
+                            vdata[i] = new Variant((sbyte)barray[index]);
                             break;
                         }
